Movie_Project/Movie_Project_Ph2/requirements.txt
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.4.22
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
greenlet==3.5.0
idna==3.13
iniconfig==2.3.0
kiwisolver==1.5.0
matplotlib==3.10.9
numpy==2.4.4
packaging==26.2
pillow==12.2.0
pluggy==1.6.0
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
RapidFuzz==3.14.5
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
thefuzz==0.22.1
typing_extensions==4.15.0
urllib3==2.6.3
